@@ -169,7 +169,7 @@ func setupReconcilers(mgr manager.Manager, cfg *reconcilerConfig) error {
 		&v1.Pod{},
 		"spec.nodeName",
 		func(o client.Object) []string {
-			pod := o.(*v1.Pod) // nolint:forcetypeassert
+			pod := o.(*v1.Pod) //nolint:forcetypeassert
 			return []string{pod.Spec.NodeName}
 		})
 	if err != nil {
